refactor: move components CartItem and ShopItems into their own files
